Extract RopeOmni handling to the function RopeOmniSilenced
I found out that RopeOmni requires the findstart call.  If that call is
executed, I think, there are less errors coming back.  It probably does
some caching, too.

You can request the RopeOmni column from the ropeCombiner by passing -1
for findstarter_index.  Usually it is the same as the ones returned from
completions in localcomplete.vim, but with the relatively recent version
that I tested it always returned zero inside of comments,  which made
the Popup Menu go crazy.