Add (inactive) functions that yield single haystack lines
Really early on, the haystack was supposed to be just a few lines and so
joining lines and a single needle.findall() call was okay.  But now this
can also be used to scan the whole buffer to replace Vim's keyword
completion.  Scanning individual lines might be a bit more frugal when
it comes to memory.