Make the case in dictionary lookup configurable
Essentially, this allows you to complete long words at the beginning of a
sentence.  The default is still to match only case-sensitively because
uppercase names would come first which is awful when completing short
leading words.

A workaround is to copy the dictionary, resort, and set the new variable to one:

    b:LocalCompleteWantIgnoreCaseDict
    g:localcomplete#WantIgnoreCaseDict